requirements.txt
pypdf==6.10.2
pdfplumber==0.11.9
pdfminer.six==20251230
